varias cosas

+++ safetrip-backend/src/main/java/com/safetripbackend/entity/Cities.java
@@ -5,8 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
 @Entity
 @AllArgsConstructor
@@ -14,7 +12,8 @@ import java.util.List;
 public class Cities {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
+
     private String name;
     private String location;
     private String sights;
