+++ safetrip-backend/src/main/java/com/safetripbackend/entity/Users.java
@@ -5,8 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
 @Entity
 @AllArgsConstructor
@@ -14,8 +12,9 @@ import java.util.List;
 public class Users {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-    private String name; // Ejemplo de atributo de tipo cadena
-    private String email; // Ejemplo de atributo de tipo cadena
-    private String password; // Ejemplo de atributo de tipo cadena
+    private Long id;
+
+    private String name;
+    private String email;
+    private String password;
 }
